Conflicts:
	statistics.txt

+++ fotolib2/src/main/java/de/k3b/io/FileCommands.java
@@ -39,6 +39,8 @@ public class FileCommands implements  Cloneable {
     public static final int OP_COPY = 1;
     public static final int OP_MOVE = 2;
     public static final int OP_DELETE = 3;
+    public static final int OP_RENAME = 4;
+    public static final int OP_UPDATE = 5;
     private static final String EXT_SIDECAR = ".xmp";
 
     private String mLogFilePath;
@@ -52,13 +54,19 @@ public class FileCommands implements  Cloneable {
 
     public int deleteFiles(String... paths) {
         int result = 0;
-        openLogfile();
-        for(String path : paths) {
-            if (deleteFileWitSidecar(new File(path))) result++;
+        if (canProcessFile(OP_DELETE)) {
+            openLogfile();
+            for (String path : paths) {
+                if (deleteFileWitSidecar(new File(path))) result++;
+            }
+            onPostProcess(paths, result, paths.length, OP_DELETE);
+            closeLogFile();
         }
-        onPostProcess(paths, result, paths.length, OP_DELETE);
-        closeLogFile();
         return result;
+    }
+
+    protected boolean canProcessFile(int opCode) {
+        return true;
     }
 
     /**
@@ -71,33 +79,45 @@ public class FileCommands implements  Cloneable {
             File sidecar = getSidecar(file);
             if (osFileExists(sidecar)) {
                 osDeleteFile(sidecar); // dont care if delete was successfull
-                log("del \"", sidecar.getAbsolutePath(), "\"");
+                log("del " , getFilenameForLog(sidecar));
             }
 
             if (osFileExists(file)) {
                 if (!osDeleteFile(file)) {
-                    log("rem file exists. delete failed : ", file.getAbsolutePath());
+                    log("rem file exists. delete failed : ", getFilenameForLog(file));
                 } else {
                     result = true; // was deleted
                 }
             } else {
-                log("rem file ", file.getAbsolutePath(), "does not exist");
+                log("rem file ", getFilenameForLog(file), " does not exist");
                 result = true; // it is gone
             }
         }
-        log("del \"", file.getAbsolutePath(), "\"");
+        log("del ", getFilenameForLog(file));
         return result;
+    }
+
+    public static String getFilenameForLog(File file) {
+        if (file == null) return "";
+        return getFilenameForLog(file.getAbsolutePath());
+    }
+
+    public static String getFilenameForLog(String absolutePath) {
+        if (absolutePath == null) return "";
+        return "\"" + absolutePath.replace("/", "\\") + "\"";
     }
 
     public int moveOrCopyFilesTo(boolean move, File destDirFolder, File... sourceFiles) {
         int result = 0;
-        if (osCreateDirIfNeccessary(destDirFolder)) {
-            File[] destFiles = createDestFiles(destDirFolder, sourceFiles);
+        if (canProcessFile(move ? OP_MOVE : OP_COPY)) {
+            if (osCreateDirIfNeccessary(destDirFolder)) {
+                File[] destFiles = createDestFiles(destDirFolder, sourceFiles);
 
-            result = moveOrCopyFiles(move, destFiles, sourceFiles);
+                result = moveOrCopyFiles(move, destFiles, sourceFiles);
 
-        } else {
-            log("rem Target dir '", destDirFolder.getAbsolutePath(), "' cannot be created");
+            } else {
+                log("rem Target dir ", getFilenameForLog(destDirFolder), " cannot be created");
+            }
         }
         return result;
     }
@@ -117,13 +137,13 @@ public class FileCommands implements  Cloneable {
 
             File destRenamed = renameDuplicate(destFile);
             if (osFileMoveOrCopy(move, destRenamed, sourceFile)) itemCount++;
-            log(((move) ? "move \"" : "copy \""), sourceFile.getAbsolutePath(), "\" \"" , destRenamed.getAbsolutePath(), "\"");
+            log(((move) ? "MOVE /y " : "COPY /y "), getFilenameForLog(sourceFile), " " , getFilenameForLog(destRenamed));
 
             File sourceSidecar = getSidecar(sourceFile);
             if (osFileExists(sourceSidecar)) {
                 File destSidecar = getSidecar(destRenamed);
                 if (osFileMoveOrCopy(move, destSidecar, sourceSidecar)) itemCount++;
-                log(((move) ? "move \"" : "copy \""), sourceSidecar.getAbsolutePath(), "\" \"" , destSidecar.getAbsolutePath(), "\"");
+                log(((move) ? "MOVE /y " : "COPY /y "), getFilenameForLog(sourceSidecar), " " , getFilenameForLog(destSidecar));
             }
             pos++;
         }
@@ -203,6 +223,7 @@ public class FileCommands implements  Cloneable {
             return file;
         }
 
+
         String filename = file.getAbsolutePath();
         String extension = ")";
         int extensionPosition = filename.lastIndexOf(".");
@@ -216,7 +237,7 @@ public class FileCommands implements  Cloneable {
             String candidatePath = filename + id + extension;
             File candidate = new File(candidatePath);
             if (!fileOrSidecarExists(candidate)) {
-                log("rem renamed from '", file.toString(), "' to '", candidatePath, "'");
+                log("rem renamed from ", getFilenameForLog(file), " to ", getFilenameForLog(candidatePath));
                 return candidate;
             }
